refactor(api): rename 'Jobs' service to 'Job'
