+++ src/api/models/job.model.js
@@ -3,7 +3,7 @@ const path = require('path')
 
 module.exports = function (app) {
   const Model = new NeDB({
-    filename: path.join(app.get('dataLocation'), 'jobs.db'),
+    filename: path.join(app.get('dataLocation'), 'job.db'),
     autoload: true
   })
