+++ src/api/services/index.js
@@ -1,7 +1,7 @@
 const movies = require('./movies/movies.service.js')
 const mediaFile = require('./media-file/media-file.service.js')
 
-const jobs = require('./jobs/jobs.service.js')
+const job = require('./job/job.service.js')
 
 const image = require('./image/image.service.js')
 
@@ -15,7 +15,7 @@ module.exports = function () {
   const app = this // eslint-disable-line no-unused-vars
   app.configure(movies)
   app.configure(mediaFile)
-  app.configure(jobs)
+  app.configure(job)
   app.configure(image)
   app.configure(scrape)
   app.configure(webhook)
