+++ src/api/services/job/job.service.js
@@ -1,7 +1,7 @@
 // Initializes the `Jobs` service on path `/jobs`
-const createService = require('feathers-nedb')
-const createModel = require('../../models/jobs.model')
-const hooks = require('./jobs.hooks')
+const createService = require('./job.class')
+const createModel = require('../../models/job.model')
+const hooks = require('./job.hooks')
 
 module.exports = function () {
   const app = this
@@ -9,16 +9,16 @@ module.exports = function () {
   const paginate = app.get('paginate')
 
   const options = {
-    name: 'jobs',
+    name: 'job',
     Model,
     paginate
   }
 
   // Initialize our service with any options it requires
-  app.use('/jobs', createService(options))
+  app.use('/job', createService(options))
 
   // Get our initialized service so that we can register hooks and filters
-  const service = app.service('jobs')
+  const service = app.service('job')
 
   service.hooks(hooks)
 }
