+++ test/api/services/job.test.js
@@ -2,9 +2,9 @@
 const assert = require('assert')
 const app = require('../../../src/api/app')
 
-describe('\'Jobs\' service', () => {
+describe('\'Job\' service', () => {
   it('registered the service', () => {
-    const service = app.service('jobs')
+    const service = app.service('job')
 
     assert.ok(service, 'Registered the service')
   })
